backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
docling>=2.0.0
python-docx
pandas
tabulate
langchain>=0.2.0
langchain-groq>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.1
